requirements.txt
# Libraries to be used in GCP.
bs4
python-dotenv